refactor(types): move sources wire types (Source, SourceStatus) to origin-types
Source and SourceStatus only depend on serde + std::path, matching
origin-types' lightweight constraint. origin-core re-exports them from
origin-types unchanged; app/src/sources/mod.rs now explicitly names the
origin-types vs origin-core import layers.
